refactor: merge api-v2 types into main api.ts file
- Move all new API types to existing api.ts file
- Remove separate api-v2.ts file
- Update imports in vector store endpoints
- Maintain single source of truth for API types
