+++ .claude-flow/metrics/performance.json
@@ -1,5 +1,5 @@
 {
-  "startTime": 1754538551388,
+  "startTime": 1754538700116,
   "totalTasks": 1,
   "successfulTasks": 1,
   "failedTasks": 0,
