+++ .claude-flow/metrics/system-metrics.json
@@ -574,5 +574,65 @@
     "cpuLoad": 0.015,
     "platform": "linux",
     "uptime": 1505.75
+  },
+  {
+    "timestamp": 1754538584585,
+    "memoryTotal": 18864345088,
+    "memoryUsed": 4661895168,
+    "memoryFree": 14202449920,
+    "memoryUsagePercent": 24.71273265121474,
+    "memoryEfficiency": 75.28726734878526,
+    "cpuCount": 6,
+    "cpuLoad": 0.05666666666666667,
+    "platform": "linux",
+    "uptime": 1535.78
+  },
+  {
+    "timestamp": 1754538614605,
+    "memoryTotal": 18864345088,
+    "memoryUsed": 4746407936,
+    "memoryFree": 14117937152,
+    "memoryUsagePercent": 25.160735312350113,
+    "memoryEfficiency": 74.83926468764989,
+    "cpuCount": 6,
+    "cpuLoad": 0.045000000000000005,
+    "platform": "linux",
+    "uptime": 1565.8
+  },
+  {
+    "timestamp": 1754538644610,
+    "memoryTotal": 18864345088,
+    "memoryUsed": 4655255552,
+    "memoryFree": 14209089536,
+    "memoryUsagePercent": 24.677536009247966,
+    "memoryEfficiency": 75.32246399075203,
+    "cpuCount": 6,
+    "cpuLoad": 0.09666666666666666,
+    "platform": "linux",
+    "uptime": 1595.8
+  },
+  {
+    "timestamp": 1754538674618,
+    "memoryTotal": 18864345088,
+    "memoryUsed": 4626808832,
+    "memoryFree": 14237536256,
+    "memoryUsagePercent": 24.526739785645717,
+    "memoryEfficiency": 75.47326021435428,
+    "cpuCount": 6,
+    "cpuLoad": 0.07833333333333332,
+    "platform": "linux",
+    "uptime": 1625.81
+  },
+  {
+    "timestamp": 1754538704635,
+    "memoryTotal": 18864345088,
+    "memoryUsed": 5068537856,
+    "memoryFree": 13795807232,
+    "memoryUsagePercent": 26.868347840096508,
+    "memoryEfficiency": 73.1316521599035,
+    "cpuCount": 6,
+    "cpuLoad": 0.16666666666666666,
+    "platform": "linux",
+    "uptime": 1655.83
   }
 ]
