+++ .claude-flow/metrics/task-metrics.json
@@ -1,10 +1,10 @@
 [
   {
-    "id": "cmd-hooks-1754538551431",
+    "id": "cmd-hooks-1754538700162",
     "type": "hooks",
     "success": true,
-    "duration": 5.760323999999997,
-    "timestamp": 1754538551437,
+    "duration": 7.495999999999981,
+    "timestamp": 1754538700170,
     "metadata": {}
   }
 ]
