+++ apps/web/src/app/api/vector-stores/[id]/route.ts
@@ -12,7 +12,7 @@ import { aiRateLimit } from '@/lib/middleware/rate-limit';
 import type {
   UpdateVectorStoreRequest,
   VectorStore,
-} from '@/lib/types/api-v2';
+} from '@/lib/types/api';
 import {
   addSecurityHeaders,
   notFoundResponse,
